fix: fix password type
