+++ packages/kit/src/components/Password/components/PasswordSetup.tsx
@@ -13,7 +13,7 @@ import {
   Unspaced,
   useForm,
 } from '@onekeyhq/components';
-import { EPasswordMode } from '@onekeyhq/kit-bg/src/services/ServicePassword/types';
+import { EPasswordMode } from '@onekeyhq/shared/types/password';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 import timerUtils from '@onekeyhq/shared/src/utils/timerUtils';
