+++ packages/kit/src/components/Password/container/PasswordSetupContainer.tsx
@@ -11,7 +11,7 @@ import {
   XStack,
 } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { EPasswordMode } from '@onekeyhq/kit-bg/src/services/ServicePassword/types';
+import { EPasswordMode } from '@onekeyhq/shared/types/password';
 import { useSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import {
   usePasswordBiologyAuthInfoAtom,
