+++ packages/kit/src/components/Password/container/PasswordVerifyPromptMount.tsx
@@ -6,7 +6,7 @@ import { useIntl } from 'react-intl';
 import { Dialog, Spinner } from '@onekeyhq/components';
 import type { IDialogShowProps } from '@onekeyhq/components/src/composite/Dialog/type';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { EPasswordPromptType } from '@onekeyhq/kit-bg/src/services/ServicePassword/types';
+import { EPasswordPromptType } from '@onekeyhq/shared/types/password';
 import { usePasswordPromptPromiseTriggerAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms/password';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
